human-factors-platform/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.3
anthropic>=0.18.1
openai>=1.12.0
opencv-python>=4.9.0.80
numpy>=1.26.3
torch>=2.1.0
torchvision>=0.16.0
pillow>=10.2.0
python-dotenv>=1.0.0
